insufficient privileges on VC creation logged to Sentry (#7144)
